# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  dto.InsightResponse:
    properties:
      account:
        type: string
      category:
        type: string
      change:
        type: number
      changePercent:
        type: number
      comparisonPeriod:
        type: string
      gainContribution:
        type: number
      href:
        type: string
      id:
        type: string
      investmentContribution:
        type: number
      period:
        type: string
      previousValue:
        type: number
      relatedAccounts:
        items:
          type: string
        type: array
      score:
        type: integer
      severity:
        type: string
      type:
        type: string
      value:
        type: number
    type: object
  dto.InsightsResponse:
    properties:
      asOf:
        type: string
      comparisonPeriod:
        type: string
      insights:
        items:
          $ref: '#/definitions/dto.InsightResponse'
        type: array
      isPartial:
        type: boolean
      period:
        type: string
    type: object
...
info:
  contact: {}
  description: Paisa is a local-first personal finance application. This API powers
    the Paisa frontend and exposes financial analysis, journal editing, configuration,
    price data, goals, transactions, and related application operations.
  title: Paisa API
  version: 0.1.0
paths:
...
  /insights:
    get:
      description: Returns deterministic financial observations and risks derived
        from ledger history
      operationId: getInsights
      parameters:
      - description: Month period in YYYY-MM format (defaults to current month)
        in: query
        name: period
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.InsightsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - PaisaAuth: []
      summary: Get deterministic financial health insights
      tags:
      - Insights
...
securityDefinitions:
  PaisaAuth:
    description: Credentials formatted as 'username:password'. Required for all /api
      endpoints when user accounts are configured in paisa.yaml.
    in: header
    name: X-Auth
    type: apiKey
swagger: "2.0"